Redirect panel has beem moved to a new file.
