+++ src/App.tsx
@@ -8,7 +8,6 @@ import {
   RouterProvider,
   Route,
   createRoutesFromElements,
-  useNavigate,
 } from 'react-router-dom'
 // this allows immer to work with Map and Set
 import { enableMapSet } from 'immer'
@@ -16,6 +15,7 @@ import { MessagePanel } from './components/Messages'
 import appConfig from './assets/config.json'
 import { KeycloakContext } from '.'
 import { useCredentialStore } from './store/CredentialStore'
+import { RedirectPanel } from './RedirectPanel'
 
 enableMapSet()
 
@@ -39,25 +39,6 @@ const routerOpts: any = {}
 
 if (appConfig.urlBaseName !== '') {
   routerOpts.basename = appConfig.urlBaseName
-}
-
-const RedirectPanel = (): JSX.Element => {
-  const navigate = useNavigate()
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      navigate('/')
-    }, 2000)
-
-    // Clean up the timer when the component is unmounted
-    return () => clearTimeout(timer)
-  }, [])
-
-  return (
-    <div>
-      <h6>Invalid URL was given. Redirecting to the application root...</h6>
-    </div>
-  )
 }
 
 const router = createBrowserRouter(
